refactor(api): move app construction into createApp factory (#624)
Security middleware + body parser + router mounts + /audit/* route
handlers + webhook raw-body pre-middleware all move from server.js into
api/createApp.js. server.js shrinks to ~180 lines of bootstrap:
arg parsing, service construction (AuthService, WikiAgentManager,
WebSocketManager, WebhookHandler), app.listen(), signal handlers.

routes/wiki is passed as an optional DI param rather than require()d
at the top of createApp, because its transitive require chain pulls
../../config/utils/config-manager.js which needs ajv resolved from the
workspace root (not installed there). Tests skip the mount; server.js
passes the real router. Filed as Vikunja #658.

Ref: docs/plans/2026-04-20-option-a-createapp-di.md
